+++ mobile/android/android-components/components/browser/storage-sync/src/main/java/mozilla/components/browser/storage/sync/PlacesHistoryStorage.kt
@@ -13,7 +13,6 @@ import androidx.work.WorkManager
 import kotlinx.coroutines.withContext
 import mozilla.appservices.places.PlacesApi
 import mozilla.appservices.places.PlacesReaderConnection
-import mozilla.appservices.places.uniffi.PlacesApiException
 import mozilla.appservices.places.uniffi.VisitObservation
 import mozilla.components.concept.base.crash.CrashReporting
 import mozilla.components.concept.storage.FrecencyThresholdOption
@@ -37,7 +36,6 @@ import mozilla.components.concept.sync.SyncStatus
 import mozilla.components.concept.sync.SyncableStore
 import mozilla.components.support.base.log.logger.Logger
 import mozilla.components.support.utils.segmentAwareDomainMatch
-import org.json.JSONObject
 
 const val AUTOCOMPLETE_SOURCE_NAME = "placesHistory"
 
@@ -283,17 +281,6 @@ open class PlacesHistoryStorage(
                 places.syncHistory(authInfo)
             }
         }
-    }
-
-    /**
-     * Import history and visits data from Fennec's browser.db file.
-     *
-     * @param dbPath Absolute path to Fennec's browser.db file.
-     * @return Migration metrics wrapped in a JSON object. See libplaces for schema details.
-     */
-    @Throws(PlacesApiException::class)
-    fun importFromFennec(dbPath: String): JSONObject {
-        return places.importVisitsFromFennec(dbPath)
     }
 
     override fun registerWithSyncManager() {